Completely reorganized into a module, tests pass but not 100% coverage
Overall, transition was not difficult.
