+++ apps/filters_data/260-ascii.txt
