+++ apps/filters_data/AF01-504_24_measured.txt
