+++ apps/filters_data/FF01-578_21_Spectrum.txt
