+++ apps/filters_data/FF596-Di01_Spectrum.txt
