+++ apps/lensviewer_app.py
@@ -1,3 +1,4 @@
+import envapp
 from mytk import *
 import raytracing as rt
 import raytracing.thorlabs as thorlabs
