+++ apps/microscope_app.py
@@ -1,3 +1,4 @@
+import envapp
 from mytk import *
 from tkinter import filedialog
 import os
