+++ apps/pydatagraph_app.py
@@ -1,3 +1,4 @@
+import envapp
 from mytk import *
 
 import os
@@ -343,7 +344,6 @@ class PyDatagraphApp(App):
         pass
 
 if __name__ == "__main__":
-    package_app_script(__file__)    
     ModulesManager.validate_environment(pip_modules={"requests":"requests","pyperclip":"pyperclip"}, ask_for_confirmation=False)
     app = PyDatagraphApp()    
     app.mainloop()
