+++ mytk/tests/.coveragerc
@@ -1,0 +1,4 @@
+[run]
+omit = 
+    config.py
+    config-3.py
