+++ mytk/tests/testBindings.py
@@ -1,3 +1,4 @@
+import envtest
 import unittest
 import os
 from mytk import *
