+++ mytk/tests/testButton.py
@@ -1,3 +1,4 @@
+import envtest
 import unittest
 from mytk import *
